requirements.txt
django
djangorestframework
web3
solana
cryptography
